order model fixed and function according to frontend

+++ backend/bolt/src/main/java/com/example/bolt/model/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String photoPath;
     private type foodType;
     private int prepareTime;
-    private float price;
+    private int price;
     private String description;
     private String restaurantID;
     private String restaurantName;
